// packages/api/src/useCases/UpdateProvider/UpdateProviderDTO.ts
export interface IUpdateProviderDTO {
  id: string
  cpfCnpj: string
  stateInscription?: string
  name: string
  company?: string
  email: string
  phone: string
  mobilePhone: string
  dateUpdated: Date
  additionalEmails?: string
  address: string
  addressNumber: string
  complement: string
  province: string
  postalCode: string
  cityId: number
  state: string
  disableAt?: Date
  observations?: string
  companyId: string
}